some refactored
